Fix for EG-341, reviewed as ENG-190, changes as follows:
This change addresses the issue raised in ENG-341.  When trying to register a ruby script through Photon, the registration process failed and the following stack-trace 
appeared in the log (partially elided):
!MESSAGE Unhandled event loop exception
!STACK 0
java.lang.VerifyError: (class: org/marketcetera/strategy/StrategyModule, method: cancelAllMarketDataRequests signature:
()V) Attempt to split long or double on the stack

The source of the problem is in JRuby which has a bug in its bindings of scalar types > 32 bits (includes doubles and floats, but not BigDecimals or int) if the types in 
question are used in the method signature of a method which was originally declared in an interface (what?!?! are you kidding?).  The problem is fixed in a later version of 
JRuby, but we cannot upgrade from 1.1.3 without some significant changes in how Ruby scripts are invoked.  We cannot make those changes now, so, instead, the fix I 
implemented was to remove the use of Java "long" types for request identifiers.  Instead, I used "int" types.  The effective difference is that users are limited to making 
2,147,483,647 simultaneous data requests across all strategies in a single JVM session instead of 9,223,372,036,854,775,807.  I strongly suspect that this will be adequate.



git-svn-id: file:///space/source.marketcetera.org/root/trunk@10192 501a4a96-914b-0410-8454-f5988a4ca596

+++ source/modules/strategy/src/main/java/org/marketcetera/strategy/OutboundServicesProvider.java
@@ -70,10 +70,10 @@ interface OutboundServicesProvider
      * 
      * @param inRequest a <code>MarketDataRequest</code> value indicating what data to request
      * @param inSource a <code>String</code> value indicating what market data provider from which to request the data
-     * @return a <code>long</code> value containing an identifier corresponding to this market data request
+     * @return an <code>int</code> value containing an identifier corresponding to this market data request
      */
-    long requestMarketData(MarketDataRequest inRequest,
-                           String inSource);
+    int requestMarketData(MarketDataRequest inRequest,
+                          String inSource);
     /**
      * Creates a market data request processed by a complex event processor. 
      *
@@ -85,22 +85,22 @@ interface OutboundServicesProvider
      * @param inStatements a <code>String[]</code> value indicating what data to request
      * @param inSource a <code>String</code> value indicating what complex event processor provider from which to request the data
      * @param inNamespace a <code>String</code> value indicating what complex event processor namespace to use
-     * @return a <code>long</code> value containing the request handle or 0 if the request failed
+     * @return an <code>int</code> value containing the request handle or 0 if the request failed
      */
-    long requestProcessedMarketData(MarketDataRequest inRequest,
-                                    String inMarketDataSource,
-                                    String[] inStatements,
-                                    String inCEPSource,
-                                    String inNamespace);
+    int requestProcessedMarketData(MarketDataRequest inRequest,
+                                   String inMarketDataSource,
+                                   String[] inStatements,
+                                   String inCEPSource,
+                                   String inNamespace);
     /**
      * Cancels a given market data request.
      * 
      * <p>If the given <code>inDataRequestID</code> identifier does not correspond to an active market data
      * request, this method does nothing.
      *
-     * @param inDataRequestID a <code>long</code> value identifying the market data request to cancel
+     * @param inDataRequestID an <code>int</code> value identifying the market data request to cancel
      */
-    void cancelMarketDataRequest(long inDataRequestID);
+    void cancelMarketDataRequest(int inDataRequestID);
     /**
      * Cancels all market data requests for this strategy.
      *
@@ -116,20 +116,20 @@ interface OutboundServicesProvider
      * @param inStatements a <code>String[]</code> value indicating what data to request
      * @param inSource a <code>String</code> value indicating what complex event processor provider from which to request the data
      * @param inNamespace a <code>String</code> value indicating what complex event processor namespace to use
-     * @return a <code>long</code> value containing an identifier corresponding to this complex event processor request
+     * @return an <code>int</code> value containing an identifier corresponding to this complex event processor request
      */
-    long requestCEPData(String[] inStatements,
-                        String inSource,
-                        String inNamespace);
+    int requestCEPData(String[] inStatements,
+                       String inSource,
+                       String inNamespace);
     /**
      * Cancels a given complex event processor data request.
      * 
      * <p>If the given <code>inDataRequestID</code> identifier does not correspond to an active complex event processor data
      * request, this method does nothing.
      *
-     * @param inDataRequestID a <code>long</code> value identifying the complex event processor request to cancel
+     * @param inDataRequestID an <code>int</code> value identifying the complex event processor request to cancel
      */
-    void cancelCEPRequest(long inDataRequestID);
+    void cancelCEPRequest(int inDataRequestID);
     /**
      * Cancels all complex event processor requests for this strategy.
      *
